content/base/test/test_bug455853.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=455853
-->
<head>
  <title>Test for Bug 455853</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=455853">Mozilla Bug 455853</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 455853 **/

function runTest() {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  var doc = document.implementation.createDocument("", "", null);
  var root = doc.createElement('nodes');
  doc.appendChild(root);
  root.expando = "ok";
  root = null;
  window.QueryInterface(Components.interfaces.nsIInterfaceRequestor)
        .getInterface(Components.interfaces.nsIDOMWindowUtils)
        .garbageCollect();
  ok(doc.documentElement.expando, "Should have preserved the expando!");
}

runTest();

</script>
</pre>
</body>
</html>
